chore: formatting

+++ src/interfaces/ILayerZeroEndpoint.sol
@@ -33,24 +33,18 @@ interface ILayerZeroEndpoint is ILayerZeroUserApplicationConfig {
         bytes calldata _srcAddress,
         address _dstAddress,
         uint64 _nonce,
-        uint _gasLimit,
+        uint256 _gasLimit,
         bytes calldata _payload
     ) external;
 
     // @notice get the inboundNonce of a lzApp from a source chain which could be EVM or non-EVM chain
     // @param _srcChainId - the source chain identifier
     // @param _srcAddress - the source chain contract address
-    function getInboundNonce(
-        uint16 _srcChainId,
-        bytes calldata _srcAddress
-    ) external view returns (uint64);
+    function getInboundNonce(uint16 _srcChainId, bytes calldata _srcAddress) external view returns (uint64);
 
     // @notice get the outboundNonce from this source chain which, consequently, is always an EVM
     // @param _srcAddress - the source chain contract address
-    function getOutboundNonce(
-        uint16 _dstChainId,
-        address _srcAddress
-    ) external view returns (uint64);
+    function getOutboundNonce(uint16 _dstChainId, address _srcAddress) external view returns (uint64);
 
     // @notice gets a quote in source native gas, for the amount that send() requires to pay for message delivery
     // @param _dstChainId - the destination chain identifier
@@ -64,7 +58,7 @@ interface ILayerZeroEndpoint is ILayerZeroUserApplicationConfig {
         bytes calldata _payload,
         bool _payInZRO,
         bytes calldata _adapterParam
-    ) external view returns (uint nativeFee, uint zroFee);
+    ) external view returns (uint256 nativeFee, uint256 zroFee);
 
     // @notice get this Endpoint's immutable source identifier
     function getChainId() external view returns (uint16);
@@ -73,31 +67,20 @@ interface ILayerZeroEndpoint is ILayerZeroUserApplicationConfig {
     // @param _srcChainId - the source chain identifier
     // @param _srcAddress - the source chain contract address
     // @param _payload - the payload to be retried
-    function retryPayload(
-        uint16 _srcChainId,
-        bytes calldata _srcAddress,
-        bytes calldata _payload
-    ) external;
+    function retryPayload(uint16 _srcChainId, bytes calldata _srcAddress, bytes calldata _payload) external;
 
     // @notice query if any STORED payload (message blocking) at the endpoint.
     // @param _srcChainId - the source chain identifier
     // @param _srcAddress - the source chain contract address
-    function hasStoredPayload(
-        uint16 _srcChainId,
-        bytes calldata _srcAddress
-    ) external view returns (bool);
+    function hasStoredPayload(uint16 _srcChainId, bytes calldata _srcAddress) external view returns (bool);
 
     // @notice query if the _libraryAddress is valid for sending msgs.
     // @param _userApplication - the user app address on this EVM chain
-    function getSendLibraryAddress(
-        address _userApplication
-    ) external view returns (address);
+    function getSendLibraryAddress(address _userApplication) external view returns (address);
 
     // @notice query if the _libraryAddress is valid for receiving msgs.
     // @param _userApplication - the user app address on this EVM chain
-    function getReceiveLibraryAddress(
-        address _userApplication
-    ) external view returns (address);
+    function getReceiveLibraryAddress(address _userApplication) external view returns (address);
 
     // @notice query if the non-reentrancy guard for send() is on
     // @return true if the guard is on. false otherwise
@@ -112,22 +95,16 @@ interface ILayerZeroEndpoint is ILayerZeroUserApplicationConfig {
     // @param _chainId - the chainId for the pending config change
     // @param _userApplication - the contract address of the user application
     // @param _configType - type of configuration. every messaging library has its own convention.
-    function getConfig(
-        uint16 _version,
-        uint16 _chainId,
-        address _userApplication,
-        uint _configType
-    ) external view returns (bytes memory);
+    function getConfig(uint16 _version, uint16 _chainId, address _userApplication, uint256 _configType)
+        external
+        view
+        returns (bytes memory);
 
     // @notice get the send() LayerZero messaging library version
     // @param _userApplication - the contract address of the user application
-    function getSendVersion(
-        address _userApplication
-    ) external view returns (uint16);
+    function getSendVersion(address _userApplication) external view returns (uint16);
 
     // @notice get the lzReceive() LayerZero messaging library version
     // @param _userApplication - the contract address of the user application
-    function getReceiveVersion(
-        address _userApplication
-    ) external view returns (uint16);
+    function getReceiveVersion(address _userApplication) external view returns (uint16);
 }
